refactor(website): one export per file for PipelineVisual types and data
Move the deck and lane data into PipelineVisual/data/ and every exported
type into PipelineVisual/types/, one export per file. Split deal-types.ts
and the pipeline-move-card type exports accordingly; pipeline-move-card.ts
now holds only movePipelineCard. Non-exported types (DragState,
FieldIconComponent) stay co-located with their single user.
