+++ packages/twenty-website/src/sections/product-feature/PipelineVisual/data/initial-lanes.ts
@@ -1,0 +1,7 @@
+import { type PipelineLanes } from '../types/pipeline-lanes';
+
+export const INITIAL_LANES: PipelineLanes = [
+  ['github', 'notion'],
+  ['airbnb'],
+  ['figma', 'stripe'],
+];
