+++ packages/twenty-website/src/sections/product-feature/PipelineVisual/data/lanes-meta.ts
@@ -1,0 +1,7 @@
+import { type CellSelectColor } from '@/app-preview/types';
+
+export const LANES_META: { color: CellSelectColor; label: string }[] = [
+  { color: 'pink', label: 'New' },
+  { color: 'purple', label: 'Meeting' },
+  { color: 'blue', label: 'Customer' },
+];
