+++ packages/twenty-website/src/sections/product-feature/PipelineVisual/types/deal-actor.ts
@@ -1,0 +1,5 @@
+export type DealActor = {
+  avatarUrl?: string;
+  name: string;
+  source: 'member' | 'system';
+};
