+++ packages/twenty-website/src/sections/product-feature/PipelineVisual/types/deal-company.ts
@@ -1,0 +1,4 @@
+export type DealCompany = {
+  domain: string;
+  name: string;
+};
