+++ packages/twenty-website/src/sections/product-feature/PipelineVisual/types/deal-person.ts
@@ -1,0 +1,4 @@
+export type DealPerson = {
+  avatarUrl: string;
+  name: string;
+};
