+++ packages/twenty-website/src/sections/product-feature/PipelineVisual/types/pipeline-card-animations.ts
@@ -1,0 +1,3 @@
+import { type PipelineCardId } from './pipeline-card-id';
+
+export type PipelineCardAnimations = Partial<Record<PipelineCardId, Animation>>;
