+++ packages/twenty-website/src/sections/product-feature/PipelineVisual/types/pipeline-card-elements.ts
@@ -1,0 +1,5 @@
+import { type PipelineCardId } from './pipeline-card-id';
+
+export type PipelineCardElements = Partial<
+  Record<PipelineCardId, HTMLDivElement | null>
+>;
