+++ packages/twenty-website/src/sections/product-feature/PipelineVisual/types/pipeline-card-id.ts
@@ -1,0 +1,6 @@
+export type PipelineCardId =
+  | 'airbnb'
+  | 'figma'
+  | 'github'
+  | 'notion'
+  | 'stripe';
