+++ packages/twenty-website/src/sections/product-feature/PipelineVisual/types/pipeline-card-rects.ts
@@ -1,0 +1,3 @@
+import { type PipelineCardId } from './pipeline-card-id';
+
+export type PipelineCardRects = Partial<Record<PipelineCardId, DOMRect>>;
