+++ packages/twenty-website/src/sections/product-feature/PipelineVisual/types/pipeline-lane-index.ts
@@ -1,0 +1,1 @@
+export type PipelineLaneIndex = 0 | 1 | 2;
