+++ packages/twenty-website/src/sections/product-feature/PipelineVisual/types/pipeline-lanes.ts
@@ -1,0 +1,7 @@
+import { type PipelineCardId } from './pipeline-card-id';
+
+export type PipelineLanes = [
+  PipelineCardId[],
+  PipelineCardId[],
+  PipelineCardId[],
+];
